Added Configs A1SR10, A1Rn10 - JeenNode V6, ATMega328P + RFM12 RFM12 - Compatibility with Configuration

+++ Source/HAL/STM32/Driver/hal_uart.c
@@ -293,7 +293,7 @@ bool hal_uart_datardy(uint8_t port)
 
 uint8_t hal_uart_get(uint8_t port)
 {
-    uint16_t retval;
+    uint8_t retval;
 
     switch(port)
     {
